repair font-related segfault, see http://bugs.debian.org/647889
Before this change, xgraph was unusable for me:
(because I set e.g., MALLOC_PERTURB_=39 in my environment)

    $ echo 1 2| MALLOC_PERTURB_=39 xgraph
    zsh: done                echo 1 2 |
    zsh: segmentation fault  xgraph

Given a bogus or unresolvable label or title font name specified with
"-size", like "helvetica-9", this code would determine that there are
no matching fonts, hence never run the body of the for loop, and then
test *font_info, which would be uninitialized stack data.  If that
happened to be nonzero, it would return 1 and the caller would
dereference that font_info pointer.
* params.c (do_font): Do not return uninitialized *font_info as if it
were a valid pointer.